Rename new CFG to main CFG creator
